[Bronze II] Title: 벌집, Time: 108 ms, Memory: 14392 KB -BaekjoonHub

+++ 백준/Bronze/2292. 벌집/벌집.java
@@ -5,7 +5,7 @@ import java.io.InputStreamReader;
 public class Main {
     public static void main(String[] args) throws IOException {
        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-       long num = Integer.parseInt(br.readLine());
+       long num = Long.parseLong(br.readLine());
         System.out.println(beeAlgorithm(num));
     }
